refactor(folders): modified folder structure again
