+++ src/presentation/ui/_layouts/AuthLayout.vue
