+++ src/presentation/ui/views/App.vue
@@ -9,5 +9,5 @@ export default {
 </script>
 
 <style lang="scss">
-@import "../assets/styles/main";
+@import "../../../assets/styles/main";
 </style>
