+++ src/presentation/utils/randomString.js
