+++ src/presentation/utils/textUtils.js
